Moved the scipy cov comparison into a test, moved graphics tests into a new file
